wip: changed SnoreView to other file
